Started work on local database, moved login and registration models
