+++ FrontEndTests/FrontEndTests.csproj
@@ -11,6 +11,7 @@
     <ItemGroup>
         <PackageReference Include="bunit.web" Version="1.39.5" />
         <PackageReference Include="Google.Cloud.Translate.V3" Version="3.9.0" />
+        <PackageReference Include="IndexedDB.Blazor.Net8" Version="1.2.1" />
         <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.4" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
         <PackageReference Include="Microsoft.Playwright" Version="1.51.0" />
