+++ SoftwareProject.Client/Models/IndexedTopicModel.cs
@@ -1,0 +1,6 @@
+﻿namespace SoftwareProject.Client.Models;
+
+public class IndexedTopicModel
+{
+    
+}
