+++ SoftwareProject.Client/Pages/Chat.razor.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Components.Web;
 using Microsoft.VisualBasic;
 using MudBlazor;
 using SoftwareProject.Client.Data;
+using SoftwareProject.Client.Providers;
 using SoftwareProject.Client.Themes;
 
 namespace SoftwareProject.Client.Pages;
@@ -15,12 +16,13 @@ public partial class Chat : ComponentBase
 {
     private List<Message> apiResponses = new();
     private bool UseFake { get; set; }
-
+    private bool isLocal { get; set; }
+    private Variant localVariant => isLocal ? Variant.Filled : Variant.Outlined;
+    private Variant onlineVariant => isLocal ? Variant.Outlined : Variant.Filled;
     private Language Language { get; set; } = Languages.English;
-    
     private bool SummariseText { get; set; }
     private string SummariseTextLabel => SummariseText ? Icons.Material.Filled.Check : Icons.Material.Filled.Close;
-    
+
     private Color SummariseButtonColor => SummariseText ? Color.Primary : Color.Default;
 
     private List<string> Fonts = new List<string>()
@@ -30,13 +32,14 @@ public partial class Chat : ComponentBase
         "Arial",
         "sans-serif"
     };
-    
+
     private string DropdownIcon =>
         QuickSettingsUp ? Icons.Material.Filled.ArrowDropUp : Icons.Material.Filled.ArrowDropDown;
+
     private bool QuickSettingsUp { get; set; }
     private bool SendingDisabled { get; set; }
     private string Question { get; set; }
-    
+
     private bool _drawerOpen = true;
     private bool _isDarkMode = true;
     private MudTheme? _theme = null;
@@ -65,27 +68,31 @@ public partial class Chat : ComponentBase
             await Submit();
         }
     }
-    
+
     private async Task Submit()
     {
         if (Question == "") return;
         string tempQuestion = (SummariseText ? "SUMMARISE THIS TEXT: " : "") + Question;
-        var translation = await http.PostAsJsonAsync("api/translate", new TranslateRequest
-            {
-                Text = tempQuestion,
-                Language = Language.Code
-            }
-        );
-        var result = await translation.Content.ReadAsStringAsync();
-        
-    apiResponses.Add(new Message
+        string translatedQuestion = "";
+        SendingDisabled = true;
+        apiResponses.Add(new Message
         {
             content = tempQuestion,
             isUser = true
         });
-        SendingDisabled = true;
+        if (Language != Languages.English)
+        {
+            var translation = await http.PostAsJsonAsync("api/translate", new TranslateRequest
+                {
+                    Text = tempQuestion,
+                    Language = Language.Code
+                }
+            );
+            translatedQuestion = await translation.Content.ReadAsStringAsync();
+            tempQuestion = translatedQuestion;
+        }
         Question = "";
-        string response = await ai.GetMessage(result, UseFake);
+        string response = await ai.GetMessage(tempQuestion, UseFake);
         apiResponses.Add(new Message
         {
             content = response,
@@ -123,8 +130,12 @@ public partial class Chat : ComponentBase
         var logoutattempt = await http.PostAsync("api/authentication/logout", null);
         if (logoutattempt.IsSuccessStatusCode)
         {
-            Snackbar.Add("You have been logged out.", Severity.Success);
-            NavigationManager.NavigateTo("/");
+            if (AuthStateProvider is CookieAuthStateProvider customProvider)
+            {
+                customProvider.NotifyAuthenticationStateChanged();
+                Snackbar.Add("You have been logged out.", Severity.Success);
+                NavigationManager.NavigateTo("/");
+            }
         }
     }
 }
@@ -134,5 +145,3 @@ class Message
     public string content { get; set; }
     public bool isUser { get; set; }
 }
-
-
