+++ SoftwareProject.Client/Pages/Login.razor
@@ -3,7 +3,6 @@
 @using SoftwareProject.Client.Layout
 @using SoftwareProject.Services
 @layout WelcomePageLayout
-@inject AccountService accountService
 @inject NavigationManager navigationManager
 @inject AuthenticationStateProvider AuthStateProvider
 @inject HttpClient http
