+++ SoftwareProject.Client/Program.cs
@@ -1,3 +1,4 @@
+using IndexedDB.Blazor;
 using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using Microsoft.EntityFrameworkCore;
@@ -5,6 +6,7 @@ using MudBlazor;
 using MudBlazor.Services;
 using SoftwareProject.Client.Providers;
 using SoftwareProject.Data;
+using SoftwareProject.Interfaces;
 using SoftwareProject.Services;
 
 var builder = WebAssemblyHostBuilder.CreateDefault(args);
@@ -18,11 +20,12 @@ builder.Services.AddScoped(sp => new HttpClient
 builder.Services.AddDbContextFactory<ChatbotDbContext>((DbContextOptionsBuilder options) =>
     options.UseSqlServer(builder.Configuration.GetConnectionString("ChatbotDbConnection")));
 
-builder.Services.AddTransient<AccountService>();
+builder.Services.AddScoped<IAccountService, AccountService>();
 builder.Services.AddTransient<ApiService>();
 builder.Services.AddMudMarkdownServices();
 builder.Services.AddAuthorizationCore();
 builder.Services.AddScoped<AuthenticationStateProvider, CookieAuthStateProvider>();
+builder.Services.AddSingleton<IIndexedDbFactory, IndexedDbFactory>();
 
 builder.Services.AddMudServices();
