+++ SoftwareProject.Client/SoftwareProject.Client.csproj
@@ -12,6 +12,7 @@
     <ItemGroup>
         <PackageReference Include="FakeItEasy" Version="8.3.0" />
         <PackageReference Include="Google.Cloud.Translate.V3" Version="3.9.0" />
+        <PackageReference Include="IndexedDB.Blazor.Net8" Version="1.2.1" />
         <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="9.0.4" />
         <PackageReference Include="Microsoft.AspNetCore.Components" Version="9.0.4" />
         <PackageReference Include="Microsoft.AspNetCore.Components.Authorization" Version="9.0.4" />
