+++ SoftwareProject/Components/App.razor
@@ -25,6 +25,8 @@
 <Routes @rendermode="InteractiveWebAssembly"/>
 <script src="_framework/blazor.web.js"></script>
 <script src=@Assets["_content/MudBlazor/MudBlazor.min.js"]></script>
+<script src="_content/MudBlazor/MudBlazor.min.js"></script>
+<script src="_content/MudBlazor.Markdown/MudBlazor.Markdown.min.js"></script>
 </body>
 
 </html>
