+++ UnitTests/UnitTests.csproj
@@ -12,6 +12,7 @@
         <PackageReference Include="bunit.core" Version="1.39.5" />
         <PackageReference Include="bunit.web" Version="1.39.5" />
         <PackageReference Include="Google.Cloud.Translate.V3" Version="3.9.0" />
+        <PackageReference Include="IndexedDB.Blazor.Net8" Version="1.2.1" />
         <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.4" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
         <PackageReference Include="Moq" Version="4.20.72" />
